+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added new framework crates:
   - `AudioToolbox` / `objc2-audio-toolbox`.
   - `AVFAudio` / `objc2-avf-audio`.
+  - `AVFoundation` / `objc2-av-foundation`.
   - `CoreAudioTypes` / `objc2-core-audio-types`.
   - `CoreGraphics` / `objc2-core-graphics`.
   - `CoreFoundation` / `objc2-core-foundation`.